change kvdb-cli name to kvdbctl
