+++ cmd/kvdbctl/cmd/connect/set.go
@@ -1,7 +1,7 @@
 package connect
 
 import (
-	"github.com/hollowdll/kvdb/cmd/kvdb-cli/config"
+	"github.com/hollowdll/kvdb/cmd/kvdbctl/config"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
