+++ cmd/kvdbctl/cmd/hashmap/hashmap.go
